+++ src/exif.cpp
@@ -953,7 +953,7 @@ namespace {
     long sumToLong(const Exiv2::Exifdatum& md)
     {
         long sum = 0;
-        for (int i = 0; i < md.count(); ++i) {
+        for (long i = 0; i < md.count(); ++i) {
             sum += md.toLong(i);
         }
         return sum;
